_import_siae: Move "timeit" helper function to a better location
It was placed in the "siaes" app, which made no particular sense for
this generic helper and was encouraging very dysfunctional cicrular
import scenarios.

Fix it by moving it to a more regular location.
