+++ itou/metabase/management/commands/_utils.py
@@ -17,8 +17,8 @@ from itou.metabase.management.commands._database_tables import (
     get_new_table_name,
     switch_table_atomically,
 )
-from itou.siaes.management.commands._import_siae.utils import timeit
 from itou.siaes.models import Siae
+from itou.utils.python import timeit
 
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
