+++ itou/siaes/management/commands/_import_siae/utils.py
@@ -6,8 +6,6 @@ Various helpers shared by the import_siae, import_geiq and import_ea_eatt script
 import csv
 import gzip
 import os
-from functools import wraps
-from time import time
 
 import pandas as pd
 from django.utils import timezone
@@ -20,31 +18,6 @@ from itou.utils.apis.geocoding import get_geocoding_data
 
 
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
-
-SHOW_IMPORT_SIAE_METHOD_TIMER = True
-
-
-def timeit(f):
-    """
-    Quick and dirty method timer (as a decorator).
-    Could not make it work easily with the `import_siae.Command` class.
-    Thus dirty becauses uses `print` instead of `self.log`.
-
-    Maybe later we can use this builtin timer instead:
-    https://docs.python.org/3/library/timeit.html#python-interface
-    """
-
-    @wraps(f)
-    def wrap(*args, **kw):
-        ts = time()
-        result = f(*args, **kw)
-        te = time()
-        msg = f"Method {f.__name__} took {te - ts:.2f} seconds to complete"
-        if SHOW_IMPORT_SIAE_METHOD_TIMER:
-            print(msg)
-        return result
-
-    return wrap
 
 
 def get_fluxiae_referential_filenames():
