+++ itou/siaes/management/commands/_import_siae/vue_af.py
@@ -20,8 +20,9 @@ an "siae_key" throughout the import_siae.py script code.
 from django.utils import timezone
 
 from itou.siaes.enums import SIAE_WITH_CONVENTION_KINDS
-from itou.siaes.management.commands._import_siae.utils import get_fluxiae_df, remap_columns, timeit
+from itou.siaes.management.commands._import_siae.utils import get_fluxiae_df, remap_columns
 from itou.siaes.models import SiaeFinancialAnnex
+from itou.utils.python import timeit
 from itou.utils.validators import validate_af_number
